Move shared utilities and type Supabase data access
What changed:

- Moved root lib/ into src/lib/ so shared app utilities live under the source root.

- Moved root types/ into src/types/ so generated database types follow the same convention.

- Typed the shared Supabase browser, server, and proxy clients with the generated Database type.

- Updated the ambassadors page to use the generated public_ambassadors_view row type after client typing exposed weak query inference.

- Simplified tsconfig path aliases to a single @/* -> ./src/* source-root mapping.
