--- lib/supabase/client.ts
@@ -1,6 +1,0 @@
-import { createBrowserClient } from '@supabase/ssr';
-
-export function createClient() {
-  // Create a supabase client on the browser with project's credentials
-  return createBrowserClient(process.env.NEXT_PUBLIC_SUPABASE_URL!, process.env.NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY!);
-}
