+++ src/lib/supabase/client.ts
@@ -1,0 +1,10 @@
+import { createBrowserClient } from '@supabase/ssr';
+import type { Database } from '@/types/database.types';
+
+export function createClient() {
+  // Create a supabase client on the browser with project's credentials
+  return createBrowserClient<Database>(
+    process.env.NEXT_PUBLIC_SUPABASE_URL!,
+    process.env.NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY!
+  );
+}
